essential_terms/requirements.txt
absl-py==0.1.12
astor==0.6.2
bleach==1.5.0
boto==2.48.0
boto3==1.6.18
botocore==1.9.18
bz2file==0.98
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
coverage==4.5.1
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
dill==0.2.7.1
docutils==0.14
ftfy==4.4.3
gast==0.2.0
gensim==3.4.0
green==2.12.1
grpcio==1.11.0
grpcio-tools==1.11.0
h5py==2.7.1
html5lib==1.0b8
idna==2.6
jmespath==0.9.3
Keras==2.1.5
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
msgpack-numpy==0.4.1
msgpack-python==0.5.4
murmurhash==0.28.0
mysqlclient==1.3.12
numpy==1.14.2
pandas==0.22.0
pathlib==1.0.1
plac==0.9.6
preshed==1.0.0
protobuf==3.5.2.post1
pydot==1.2.4
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
regex==2017.4.5
requests==2.18.4
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
smart-open==1.5.7
spacy==2.0.10
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
termstyle==0.1.11
thinc==6.10.2
toolz==0.9.0
tqdm==4.19.8
ujson==1.35
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
